made more adjustments to outputs

+++ readelf/0-hreadelf.c
@@ -82,16 +82,16 @@ void print_elf_header(const char *filename) {
     printf("  Type:                              %s\n", get_elf_type(ehdr.e_type));
     printf("  Machine:                           %s\n", get_machine_type(ehdr.e_machine));
     printf("  Version:                           0x%x\n", ehdr.e_version);
-    printf("  Entry point address:               0x%lx\n", ehdr.e_entry);
-    printf("  Start of program headers:          %ld (bytes into file)\n", ehdr.e_phoff);
+    printf("  Entry point address:               0x%lx\n", (unsigned long)ehdr.e_entry);
+    printf("  Start of program headers:          %lu (bytes into file)\n", (unsigned long)ehdr.e_phoff);
     printf("  Start of section headers:          %lu (bytes into file)\n", (unsigned long)ehdr.e_shoff);
     printf("  Flags:                             0x%x\n", ehdr.e_flags);
     printf("  Size of this header:               %u (bytes)\n", (unsigned int)ehdr.e_ehsize);
-    printf("  Size of program headers:           %d (bytes)\n", ehdr.e_phentsize);
-    printf("  Number of program headers:         %d\n", ehdr.e_phnum);
-    printf("  Size of section headers:           %d (bytes)\n", ehdr.e_shentsize);
-    printf("  Number of section headers:         %d\n", ehdr.e_shnum);
-    printf("  Section header string table index: %d\n", ehdr.e_shstrndx);
+    printf("  Size of program headers:           %u (bytes)\n", (unsigned int)ehdr.e_phentsize);
+    printf("  Number of program headers:         %u\n", (unsigned int)ehdr.e_phnum);
+    printf("  Size of section headers:           %u (bytes)\n", (unsigned int)ehdr.e_shentsize);
+    printf("  Number of section headers:         %u\n", (unsigned int)ehdr.e_shnum);
+    printf("  Section header string table index: %u\n", (unsigned int)ehdr.e_shstrndx);
     
     close(fd);
 }
